Implement stack with max stack size
Added names to chunks, and a stack to the engine.
The stack can (but doesn't have to) have a max stack size.
If that stack size is reached by the program, it will (lua-)error and attach the current stack to the error. The engine will return the stack error.